Refactor models.ts to separate files
