+++ src/models/API.ts
@@ -1,7 +1,4 @@
-import type { RouteData } from "./schemas/mod.ts";
-
-// Large swathes of this file are currently commented because the
-// implementation is either non-existent or untested.
+import type { RouteData } from "../schemas/mod.ts";
 
 /**
  * A low-level wrapper over the Jamf School API. It serves as a replacement
@@ -486,369 +483,3 @@ type APIMoveUserInit = {
 	/** Whether to only move the user, instead of the user and their devices. */
 	onlyUser?: boolean;
 };
-
-/**
- * A high-level abstraction over the Jamf School API that allows you to reason
- * about data as objects with methods.
- *
- * A client instance does not act on the objects it creates, only acting as a
- * way to get other objects that can perform actions.
- *
- * Lower-level actions can be done with the API, for example bulk
- * editing operations.
- */
-export interface Client {
-	/** Discriminator for type checks. */
-	readonly type: "Client";
-
-	/**
-	 * Create a device object with data from the API.
-	 *
-	 * If you don't already have the data, you may want `getDeviceById` or
-	 * `getDeviceBySerialNumber`.
-	 */
-	createDevice(data: APIData["getDevices"][number]): Device;
-
-	/**
-	 * Create a device group object with data from the API.
-	 *
-	 * If you don't already have the data, you may want `getDeviceGroupById`.
-	 */
-	createDeviceGroup(data: APIData["getDeviceGroup"]): DeviceGroup;
-
-	/**
-	 * Create a user object with data from the API.
-	 *
-	 * If you don't already have the data, you may want `getUserById`.
-	 */
-	createUser(data: APIData["getUser"]): User;
-
-	/**
-	 * Create a user group object with data from the API.
-	 *
-	 * If you don't already have the data, you may want `getUserGroupById`.
-	 */
-	createUserGroup(data: APIData["getUserGroup"]): UserGroup;
-
-	/** (Read) Get a single user by their ID. */
-	getUserById(id: number): Promise<User | null>;
-
-	/** (Read) Get all users. */
-	getUsers(): Promise<User[]>;
-
-	/** (Read) Get a single user group by its ID. */
-	getUserGroupById(id: number): Promise<UserGroup | null>;
-
-	/** (Read) Get a single user group by its name. */
-	getUserGroupByName(name: string): Promise<UserGroup | null>;
-
-	/** (Read) Get all user groups. */
-	getUserGroups(): Promise<UserGroup[]>;
-
-	/** (Read) Get a single device by its UDID. */
-	getDeviceById(udid: string): Promise<Device | null>;
-
-	/** (Read) Get a single device by its serial number. */
-	getDeviceBySerialNumber(serialNumber: string): Promise<Device | null>;
-
-	/** (Read) Get all devices. */
-	getDevices(): Promise<Device[]>;
-
-	/**
-	 * (Read) Get all devices in the given groups.
-	 *
-	 * This is a far more efficient method than calling `getDevices` on
-	 * each `DeviceGroup` and checking for uniqueness.
-	 */
-	getDevicesInGroups(deviceGroups: DeviceGroup[]): Promise<Device[]>;
-
-	/** (Read) Get a single device group by its ID. */
-	getDeviceGroupById(id: number): Promise<DeviceGroup | null>;
-
-	/** (Read) Get a single device group by its name. */
-	getDeviceGroupByName(name: string): Promise<DeviceGroup | null>;
-
-	/** (Read) Get all device groups. */
-	getDeviceGroups(): Promise<DeviceGroup[]>;
-}
-
-/**
- * Device represents a single device.
- */
-export interface Device {
-	/** Discriminator for type checks. */
-	readonly type: "Device";
-
-	/**
-	 * The unique device identifier.
-	 *
-	 * The UDID is calculated based on the device's hardware rather than being
-	 * randomly assigned during manufacturing.
-	 */
-	readonly udid: string;
-
-	/**
-	 * The device's serial number.
-	 *
-	 * Serial numbers are less unique than the UDID, but are easier for humans
-	 * to reason about and type.
-	 */
-	readonly serialNumber: string;
-
-	/** The device name. */
-	readonly name: string;
-
-	/**
-	 * The management status of the device.
-	 *
-	 * Management is a Jamf concept. Unmanaged devices are devices that haven't
-	 * check in with Jamf in 7 days.
-	 */
-	readonly isManaged: boolean;
-
-	/**
-	 * The supervision status of the device.
-	 *
-	 * Jamf has greater control over supervised devices, as they are completely
-	 * managed by Jamf. Most features enabled by supervision are not accessible
-	 * though the API.
-	 */
-	readonly isSupervised: boolean;
-
-	/** The device class is the kind of device (eg. `ipad`) */
-	readonly deviceClass: string;
-
-	/** The device's asset tag. */
-	readonly assetTag: string;
-
-	/** Human-friendly name of the operating system (eg. `"iOS 15.0.0"`) */
-	readonly os: string;
-
-	/** The operating system prefix (eg. `"iOS"`) */
-	readonly osPrefix: string;
-
-	/** The version of the operating system (eg. `"15.0.0"`) */
-	readonly osVersion: string;
-
-	/** Human-friendly model name of the device. */
-	readonly modelName: string;
-
-	/**
-	 * The model identifier is a string that uniquely identifies the specific
-	 * device model (eg. `iPad11,6`).
-	 *
-	 * Search with [your favourite search engine] for "apple model identifier
-	 * list site:github.com" and you'll find many (non-comprehensive) lists for
-	 * different kinds of devices.
-	 */
-	readonly modelIdentifier: string;
-
-	/** The name for the category of device (eg. `iPad`) */
-	readonly modelType: string;
-
-	/** The remaining charge as a percentage (number between 0 and 1) */
-	readonly batteryPercentage: number;
-
-	/**
-	 * The total capacity of the battery in watt-hours (Wh).
-	 *
-	 * Jamf does not report the battery volatage or milliamp-hours.
-	 *
-	 * You can calculate the milliamp hours if you know the battery voltage.
-	 * Most of Apple's mobile devices use ~3.8V batteries.
-	 *
-	 * ```
-	 * const mAh = device.batteryCapacity * 1000 / V
-	 * ```
-	 *
-	 * You can calculate the voltage if you know the milliamp-hours.
-	 *
-	 * ```
-	 * const V = device.batteryCapacity / mAh * 1000
-	 * ```
-	 *
-	 * NOTE: I know nothing about electricity. If this is wrong, let me know!
-	 */
-	readonly batteryCapacity: number;
-
-	/**
-	 * (Read) Update this device's data.
-	 *
-	 * Other devices created from the same data will not be updated.
-	 */
-	update(): Promise<this>;
-
-	/** (Read) Get the device's owner, if any. */
-	getOwner(): Promise<User | null>;
-
-	/**
-	 * (Edit) Assign a new owner to this device.
-	 *
-	 * The owner can be set by using an object with an `id` property (number).
-	 * ```
-	 * device.setOwner(testAccount)
-	 * ```
-	 *
-	 * Alternatively, if you know the ID of the user you want to use, an object
-	 * literal may be more appropriate.
-	 * ```
-	 * device.setOwner({ id: 4 });
-	 * ```
-	 */
-	setOwner(user: { id: number }): Promise<this>;
-
-	/** (Read) Get the device's groups. */
-	getGroups(): Promise<DeviceGroup[]>;
-
-	/** (Add) Schedule a restart. */
-	restart(): Promise<this>;
-
-	/** (Add) Schedule a complete wipe. */
-	wipe(): Promise<this>;
-}
-
-/**
- * DeviceGroup represents a named collection of devices.
- */
-export interface DeviceGroup {
-	/** Discriminator for type checks. */
-	readonly type: "DeviceGroup";
-
-	/** The Jamf-assigned ID of the device group. */
-	readonly id: number;
-
-	/** The device group's name. */
-	readonly name: string;
-
-	/**
-	 * A description of the structure or contents of the device group.
-	 * This may be empty.
-	 */
-	readonly description: string;
-
-	/** Additional relevant information. This may be empty. */
-	readonly information: string;
-
-	/** The URL used for the device group image, null if not set. */
-	readonly imageUrl: string | null;
-
-	/**
-	 * Whether the group is static (false) or smart (true).
-	 *
-	 * Smart groups select their members based on filters.
-	 */
-	readonly isSmartGroup: boolean;
-
-	// /**
-	//  * _If you know what this is, please raise an issue to tell me._
-	//  * https://github.com/SeparateRecords/deno_jamf_school/issues/new
-	//  */
-	// readonly isShared: boolean;
-
-	// /**
-	//  * _If you know what this is, please raise an issue to tell me._
-	//  * https://github.com/SeparateRecords/deno_jamf_school/issues/new
-	//  */
-	// readonly isClass: boolean;
-
-	/**
-	 * (Read) Update this device group's data.
-	 *
-	 * Other device groups created from the same data will not be updated.
-	 */
-	update(): Promise<this>;
-
-	/**
-	 * (Read) Get all the devices in the device group.
-	 *
-	 * Note that if you have an array of device groups, it is _far_ more efficient
-	 * to call the following:
-	 * ```
-	 * await client.getDevicesInGroups(array)
-	 * ```
-	 * Don't map over an array of device groups - filter them and use the method
-	 * shown above. If you see this pattern in your code, it can be replaced.
-	 * ```
-	 * await Promise.all(array.map((group) => group.getDevices()))
-	 * ```
-	 */
-	getDevices(): Promise<Device[]>;
-}
-
-/**
- * User represents a single user.
- */
-export interface User {
-	/** Discriminator for type checks. */
-	readonly type: "User";
-
-	/** The Jamf-assigned ID of the user. */
-	readonly id: number;
-
-	/** The user's email address. */
-	readonly email: string;
-
-	/** The username of the user. */
-	readonly username: string;
-
-	/** The user's domain, if they were imported from a LDAP server. */
-	readonly domain: string;
-
-	/** The user's full name. */
-	readonly name: string;
-
-	/** The user's first name.*/
-	readonly firstName: string;
-
-	/** The user's last name.*/
-	readonly lastName: string;
-
-	/** Additional notes associated with the user. */
-	readonly notes: string;
-
-	/** Whether the user is in the trash (deleted). */
-	readonly isTrashed: boolean;
-
-	/** Whether the user is excluded from teacher restrictions. */
-	readonly isExcludedFromRestrictions: boolean;
-
-	/**
-	 * (Read) Update this user's data.
-	 *
-	 * Other users created from the same data will not be updated.
-	 */
-	update(): Promise<this>;
-
-	/** (Read) Get all the devices assigned to this user. */
-	getDevices(): Promise<Device[]>;
-
-	/** (Read) Get all the groups the user is in. */
-	getGroups(): Promise<UserGroup[]>;
-}
-
-/**
- * UserGroup represents a named group of users.
- */
-export interface UserGroup {
-	/** Discriminator for type checks. */
-	readonly type: "UserGroup";
-
-	/** The ID of the user group. */
-	readonly id: number;
-
-	/** The name of the user group. */
-	readonly name: string;
-
-	/** A description of the user group (may be empty) */
-	readonly description: string;
-
-	/**
-	 * (Read) Update this user group's data.
-	 *
-	 * Other user groups created from the same data will not be updated.
-	 */
-	update(): Promise<this>;
-
-	/** (Read) Get all the users in the user group. */
-	getUsers(): Promise<User[]>;
-}
